+++ app/src/main/res/values-en/strings.xml
@@ -48,7 +48,7 @@
     <string name="menu_title_import_view_rule_description">"Select import rule image "</string>
     <string name="field_activity">Activity:%1$s</string>
     <string name="field_view">View:%1$s</string>
-    <string name="share_crash_info">Share crash info</string>
+    <string name="dialog_btn_copy">COPY</string>
     <string name="wechat_donate_tips">Select QR code scan from gallery</string>
     <string name="rule_details_field_create_time">Create time</string>
     <string name="rule_details_field_generate_version">Match version</string>
